Make scheduler stub match Bukkit API
