+++ src/main/java/org/bukkit/scheduler/BukkitScheduler.java
@@ -1,29 +1,24 @@
 package org.bukkit.scheduler;
 
-public class BukkitScheduler {
-    public BukkitTask runTaskLater(Object plugin, Runnable task, long delay) {
-        task.run();
-        return new DummyBukkitTask();
-    }
+/**
+ * Minimal scheduler interface used for testing without a real Bukkit runtime.
+ *
+ * <p>This mirrors the Bukkit API's {@code BukkitScheduler} which is an
+ * interface on modern servers. The previous stub incorrectly declared it as a
+ * class which caused code compiled against it to expect a concrete class at
+ * runtime. When running on a real server the scheduler is provided as an
+ * interface, leading to {@link java.lang.IncompatibleClassChangeError}. By
+ * modelling it as an interface here we ensure the compiled bytecode matches the
+ * actual API.</p>
+ */
+public interface BukkitScheduler {
 
-    public BukkitTask runTaskLaterAsynchronously(Object plugin, Runnable task, long delay) {
-        task.run();
-        return new DummyBukkitTask();
-    }
+    BukkitTask runTaskLater(Object plugin, Runnable task, long delay);
 
-    public int scheduleSyncRepeatingTask(Object plugin, Runnable task, long delay, long period) {
-        task.run();
-        return 0;
-    }
+    BukkitTask runTaskLaterAsynchronously(Object plugin, Runnable task, long delay);
 
-    public void cancelTask(int id) {
-        // no-op
-    }
+    int scheduleSyncRepeatingTask(Object plugin, Runnable task, long delay, long period);
 
-    private static class DummyBukkitTask implements BukkitTask {
-        @Override
-        public void cancel() {
-            // no-op
-        }
-    }
+    void cancelTask(int id);
 }
+
